Move exceptions into their own module
